III-5344 - Move eventTypesWithNoThemes to constant in EventTypes.ts
